chg ShowDistance to take doubles
git-svn-id: http://vtp.googlecode.com/svn/trunk@5144 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/Enviro/Enviro.h
@@ -201,7 +201,7 @@ public:
 	virtual void CameraChanged() {}
 	virtual void EarthPosUpdated() {}
 	virtual void ShowDistance(const DPoint2 &p1, const DPoint2 &p2,
-		float fGround, float fVertical) {}
+		double fGround, double fVertical) {}
 	virtual vtTagArray *GetInstanceFromGUI() { return NULL; }
 	virtual bool OnMouseEvent(vtMouseEvent &event) { return false; }
 	virtual void SetTimeEngineToGUI(TimeEngine *pEngine) {}
